# swagger.yaml
openapi: 3.0.0
info:
  title: OAuth2 Server API
  version: "1.0.0"
  description: API для управления OAuth2 аутентификацией и авторизацией.
servers:
  - url: http://localhost:3000
paths:
...
  /user/profile:
    get:
      summary: Получение профиля пользователя
      description: Get user profile information
      operationId: GetUserProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Успешное получение профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Неавторизованный
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
    patch:
      summary: Обновление профиля пользователя
      description: Update user profile information
      operationId: UpdateUserProfile
      security:
        - BearerAuth: []
      requestBody:
        description: Данные для обновления профиля
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: Успешное обновление профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: Неавторизованный
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
  /user/account:
    delete:
      summary: Удаление аккаунта пользователя
      description: Delete user account
      operationId: DeleteUserAccount
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Успешное удаление аккаунта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
        '401':
          description: Неавторизованный
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
  /user/backup:
    get:
      summary: Резервное копирование данных пользователя
      description: Backup user data
      operationId: BackupUserData
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Успешное резервное копирование
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupResponse'
        '401':
          description: Неавторизованный
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
  /user/restore:
    post:
      summary: Восстановление данных пользователя
      description: Restore user data
      operationId: RestoreUserData
      security:
        - BearerAuth: []
      requestBody:
        description: Данные для восстановления
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
      responses:
        '200':
          description: Успешное восстановление
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreResponse'
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: Неавторизованный
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LogoutSuccess:
      type: object
      properties:
        message:
          type: string
          example: "logged out"
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
      required:
        - id
        - email
        - name
    UpdateUserProfileRequest:
      type: object
      properties:
        email:
          type: string
          example: "new_email@example.com"
        name:
          type: string
          example: "Jane Doe"
      required:
        - email
        - name
    BackupResponse:
      type: object
      properties:
        backup_id:
          type: string
          example: "backup_123"
        status:
          type: string
          example: "completed"
    RestoreRequest:
      type: object
      properties:
        backup_id:
          type: string
          example: "backup_123"
      required:
        - backup_id
    RestoreResponse:
      type: object
      properties:
        restore_id:
          type: string
          example: "restore_123"
        status:
          type: string
          example: "completed"
    HTTPError:
      type: object
      properties:
        error:
          type: string
          example: "Неверные данные запроса"
      required:
        - error